backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
celery==5.4.0
openai==1.30.1
pinecone-client==3.2.2
google-cloud-pubsub==2.21.1
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.23.1
authlib==1.3.1
cryptography==42.0.7
python-dotenv==1.0.1
structlog==24.2.0
sentry-sdk[fastapi]==2.2.1
prometheus-fastapi-instrumentator==7.0.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
factory-boy==3.3.0
faker==25.2.0